requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
black==24.4.2
boto3==1.34.148
botocore==1.34.148
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cycler==0.12.1
flower==2.0.1
fonttools==4.53.1
gevent==24.2.1
greenlet==3.0.3
humanize==4.10.0
idna==3.7
importlib_resources==6.4.0
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.9.1
mypy-extensions==1.0.0
numpy==2.0.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
PyAudio==0.2.14
pycparser==2.22
pygame==2.6.0
pyparsing==3.1.2
pyTelegramBotAPI==4.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sounddevice==0.4.7
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
